move generate file path
